src/deepseek_ocr/preprocessing.py

from transformers import AutoModel, AutoTokenizer
import torch
import os
from PIL import Image

class DeepSeekOCRPreprocessor:
    def __init__(self, model_name='deepseek-ai/DeepSeek-OCR'):
        self.tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)        self.model = AutoModel.from_pretrained(model_name, trust_remote_code=True, use_safetensors=True)        self.model = self.model.eval().cuda().to(torch.bfloat16)

    def preprocess_text_to_vision_tokens(self, text_input, image_file=None, output_path='./output'):
        """
        Converts text and an optional image into DeepSeek-OCR's vision tokens.
        If image_file is provided, it will be used as the primary input.
        If only text_input is provided, it will be wrapped into a prompt.
        """
        if image_file:
            # The model's infer method takes an image file directly
            # The prompt can guide the OCR process, e.g., 

            # prompt = f"<image>\n<|grounding|>Convert the document to markdown. {text_input}"
            # For now, let's assume the image itself contains the primary information.
            # We might need to adjust this based on how DeepSeek-OCR handles text alongside image_file.
            # Based on the example, the prompt usually describes the action on the image.
            prompt = f"<image>\n<|grounding|>Convert the document to markdown. " # Default prompt for document conversion
            if text_input: # If there's additional text, it might be part of the prompt or ignored for image-centric tasks
                prompt += text_input # Append text to prompt if it's meant as instruction
            
            # Ensure output directory exists
            os.makedirs(output_path, exist_ok=True)

            res = self.model.infer(self.tokenizer, prompt=prompt, image_file=image_file, output_path=output_path, 
                                   base_size=1024, image_size=640, crop_mode=True, save_results=True, test_compress=True)
            # The `res` object from model.infer would contain the vision tokens or processed output.
            # For simplicity, we assume `res` directly contains or can be used to retrieve the compressed vision tokens.
            # Actual implementation might need to parse `res` to extract the specific vision token representation.
            return res # This needs to be refined based on actual DeepSeek-OCR output format
        else:
            # If no image file, we need to convert text to a visual representation first.
            # This part is conceptual as DeepSeek-OCR primarily works on images.
            # For now, we will raise an error or return a placeholder.
            raise NotImplementedError("Direct text-to-image conversion for DeepEncoder without an initial image is not directly supported by the provided DeepSeek-OCR examples. An image input is expected.")

    def decode_vision_tokens(self, vision_tokens_output):
        """
        Decodes the vision tokens back into text using the DeepSeek3B-MoE-A570M decoder.
        This method assumes the `vision_tokens_output` is the result from `preprocess_text_to_vision_tokens`
        and that `self.model.infer` already performs the decoding or returns a decodable object.
        """
        # The `model.infer` method directly returns the processed output, which can be text or other formats.
        # If `test_compress=True` is used, it often implies the output is already processed/decoded.
        # For this integration, we'll assume the result `res` from `infer` is the decoded text or a structure containing it.
        # More specific parsing might be needed based on the actual `res` structure.
        return vision_tokens_output # Placeholder, actual implementation would extract text from the result


# Example Usage (for testing purposes)
if __name__ == "__main__":
    # This part needs actual image files to run.
    # For demonstration, we'll assume a dummy image path.
    # You would replace 'path/to/your/document.jpg' with a real image file.
    # You would also need to install the DeepSeek-OCR dependencies (transformers, torch, etc.)

    # os.environ["CUDA_VISIBLE_DEVICES"] = '0' # Uncomment if you have a GPU

    preprocessor = DeepSeekOCRPreprocessor()

    # --- Example 1: Processing an image file ---
    print("\n--- Processing an image file ---")
    dummy_image_path = "./DeepSeek-OCR/assets/demo.jpg" # Assuming a demo image exists in the cloned repo
    if os.path.exists(dummy_image_path):
        try:
            # The prompt here is for the OCR task on the image.
            # The text_input parameter in preprocess_text_to_vision_tokens is for additional instructions.
            processed_output = preprocessor.preprocess_text_to_vision_tokens(
                text_input="Extract all text and convert to markdown.",
                image_file=dummy_image_path,
                output_path="./deepseek_ocr_output"
            )
            print("Processed output (vision tokens/decoded text):", processed_output)
            # If processed_output needs further decoding, call decode_vision_tokens
            # decoded_text = preprocessor.decode_vision_tokens(processed_output)
            # print("Decoded text:", decoded_text)
        except Exception as e:
            print(f"Error processing image: {e}")
    else:
        print(f"Dummy image not found at {dummy_image_path}. Please provide a valid image path to test.")

    # --- Example 2: Attempting to process text directly (will raise NotImplementedError as per current design) ---
    print("\n--- Attempting to process text directly ---")
    try:
        preprocessor.preprocess_text_to_vision_tokens(text_input="This is a long document that needs compression.")
    except NotImplementedError as e:
        print(f"Caught expected error: {e}")
    except Exception as e:
        print(f"Caught unexpected error: {e}")

