requirements.txt
google-generativeai
transformers
torch
torchvision
torchaudio
vllm==0.8.5

